Tighten household module boundaries
